# encoding/base64: optimize decodemap memory set