… src/encoding/base32/base32.go
… src/encoding/base64/base64.go